feat: support comparisons of stocks and caching
I did the a while ago, so I'm not super sure what these changes actually
were. From the looks of it, I updated the caching mechanism for storing
data (important due to the rate limiting of alphavantage's API) and
added basic stuff for plotting. I want to start working on testing some
trading algorithms, so just going to commit this stuff now.